+++ product-reservation-service/src/main/java/com/monkey/productreservationservice/presentation/controller/ProductReservationControllerApiV1.java
@@ -8,9 +8,9 @@ import com.monkey.productreservationservice.application.dto.response.ResProductR
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationGetDTOApiV1;
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationPostByIdCancelDTOApiV1;
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationPostDTOApiV1;
+import com.monkey.productreservationservice.application.service.ProductReservationServiceApiV1;
 import com.monkey.productreservationservice.domain.entity.ProductReservationEntity;
 import com.monkey.productreservationservice.domain.repository.ProductReservationRepository;
-import com.monkey.productreservationservice.domain.vo.ProductReservationStatus;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -25,25 +25,16 @@ import java.util.UUID;
 @RequestMapping("/v1/product-reservations")
 public class ProductReservationControllerApiV1 {
     private final ProductReservationRepository productReservationRepository; // 추후에 서비스로 이동
+    private final ProductReservationServiceApiV1 productReservationService;
 
-    // 예약 생성
+    // 예약 등록
     @PostMapping("/{productId}")
     public ResponseEntity<ResDTO<ResProductReservationPostDTOApiV1>> postBy(
             @PathVariable UUID productId,
             @RequestBody @Valid ReqProductReservationPostDTOApiV1 reqDto
             ) {
-        // storeId, userId -> 임시로 값 받아옴
-        UUID storeId = UUID.randomUUID();
-        long userId = 123L;
-
-        ProductReservationEntity productReservationEntity = reqDto.getProductReservation().toEntity(
-                productId, userId, storeId, ProductReservationStatus.PENDING_PICKUP
-        );
-
-        ProductReservationEntity saved = productReservationRepository.save(productReservationEntity);
-
-        ResProductReservationPostDTOApiV1 resDto = ResProductReservationPostDTOApiV1.of(saved);
-
+        long userId = 123L; // 추후에 인증된 정보에서 받아오기
+        ResProductReservationPostDTOApiV1 resDto = productReservationService.postBy(productId, reqDto, userId);
         return new ResponseEntity<>(ResDTO.success(resDto), HttpStatus.OK);
     }
 
@@ -52,21 +43,14 @@ public class ProductReservationControllerApiV1 {
     public ResponseEntity<ResDTO<ResProductReservationPostByIdCancelDTOApiV1>> cancelBy(
             @PathVariable UUID productReservationId
     ) {
-        ProductReservationEntity productReservationEntity = getActiveProductReservationById(productReservationId);
-
-        productReservationEntity.delete(123L);
-
-        ProductReservationEntity saved = productReservationRepository.save(productReservationEntity);
-
-        ResProductReservationPostByIdCancelDTOApiV1 resDto = ResProductReservationPostByIdCancelDTOApiV1.of(saved);
+        ResProductReservationPostByIdCancelDTOApiV1 resDto = productReservationService.cancelBy(productReservationId)
         return new ResponseEntity<>(ResDTO.success(resDto), HttpStatus.OK);
     }
 
     // 예약내역 전체 조회
     @GetMapping
     public ResponseEntity<ResDTO<ResProductReservationGetDTOApiV1>> getBy() {
-        List<ProductReservationEntity> productReservationList = productReservationRepository.findAllByIsDeletedFalse();
-        ResProductReservationGetDTOApiV1 resDto = ResProductReservationGetDTOApiV1.of(productReservationList);
+        ResProductReservationGetDTOApiV1 resDto = productReservationService.getBy();
         return new ResponseEntity<>(ResDTO.success(resDto), HttpStatus.OK);
     }
 
@@ -78,10 +62,6 @@ public class ProductReservationControllerApiV1 {
         return new ResponseEntity<>(ResDTO.success(resDto), HttpStatus.OK);
     }
 
-    // 존재하는 예약 검증 메서드
-    private ProductReservationEntity getActiveProductReservationById(UUID productReservationId) {
-        return productReservationRepository.findByProductReservationIdAndIsDeletedFalse(productReservationId)
-                .orElseThrow(() -> new CustomException(ResponseCode.NOT_FOUND));
-    }
+
 
 }
